// WADNR.Web/src/app/shared/generated/model/update-workflow-progress-response.ts
export class UpdateWorkflowProgressResponse { 
    ProjectUpdateBatchID?: number;
    ProjectID?: number;
    ProjectName?: string | null;
    ProjectUpdateStateID?: number;
    ProjectUpdateStateName?: string | null;
    LastUpdateDate?: string;
    LastUpdatedByPersonName?: string | null;
    SubmittedByPersonName?: string | null;
    SubmittedDate?: string | null;
    ReturnedByPersonName?: string | null;
    ReturnedDate?: string | null;
    CanSubmit?: boolean;
    IsReadyToApprove?: boolean;
    Steps?: { [key: string]: WorkflowStepStatus; } | null;
    ReviewerComments?: { [key: string]: string | null; } | null;
    CanEdit?: boolean;
    CanApprove?: boolean;
    CanReturn?: boolean;
    CanDelete?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}